perl-project/requirements.txt
Django==2.1.5
django-rest-framework==0.1.0
djangorestframework==3.9.1
pytz==2018.9
